+++ test/postdeploy.test.js
@@ -22,11 +22,11 @@ function getbaseurl() {
   const namespace = 'helix';
   const package = 'helix-services';
   const name = packjson.name.replace('@adobe/helix-', '');
-  let version = `v${packjson.version}`;
-  if (process.env.CI && process.env.CIRCLE_BUILD_NUM) {
+  let version = `${packjson.version}`;
+  if (process.env.CI && process.env.CIRCLE_BUILD_NUM && process.env.CIRCLE_BRANCH !== 'master') {
     version = `ci${process.env.CIRCLE_BUILD_NUM}`;
   }
-  return `api/v1/web/${namespace}/${package}/${name}%40${version}`;
+  return `api/v1/web/${namespace}/${package}/${name}@${version}?static.owner=trieloff&static.repo=helix-demo&static.ref=master&path=/index.md`;
 }
 
 describe('Running Post-Deployment Integration Tests', () => {